pyproject.toml
[tool.poetry]
name = "ai-law"
version = "0.1.0"
description = ""
authors = ["alumihumisang <rita86822002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
neo4j = "^5.27.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.3.1"
chainlit = "^1.3.2"
langchain-ollama = "^0.2.2"
faiss-cpu = "^1.9.0.post1"
google-api-python-client = "^2.156.0"
google-auth = "^2.37.0"
langchain = "^0.3.13"
ollama = "^0.4.5"
openai = "^1.58.1"
gspread = "^6.1.4"
pandas = "^2.2.3"
flask = "^3.1.0"
mymodule = "^1.0.0"
docx = "^0.2.4"
torch = "^2.6.0"
ace-tools = "^0.0"
tqdm = "^4.67.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
